+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [1.1.13] - 2026-02-01
+
+### Summary
+
+refactor(tests): code relationship mapping with 2 supporting modules
+
+### Core
+
+- update src/clonebox/cloner.py
+
+### Test
+
+- update tests/test_cloner.py
+- update tests/test_network.py
+
+
 ## [1.1.12] - 2026-02-01
 
 ### Summary